ui: Implement scrolling reactively
Instead of noticing every situation when we might need to scroll, keep
track of refs for the currently selected pile and card within each pile.
Then we simply scroll to the selected element when it changes.

We pass functions for the template refs, which is a little awkward. We
could pass a string and get an array ref, but Vue apparently does not
guarantee this will match our iteration order after mutation. There are
patterns[1] for storing the array in order, but we just store the
selected elements.

The array of picked cards is also awkward. This should be cleaner if we
componentize the card carousel.

The carousel currently also may cause the piles pane to scroll
vertically. Without resorting to scrolling to pixel positions, support
for scrolling just the horizontal carousel is being implemented by browsers with
the container option to scrollIntoView.[2][3]

[1]: https://www.youtube.com/watch?v=haIxgwHKi88
[2]: https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollIntoView
[3]: https://chromestatus.com/feature/5100036528275456