ui: Generalize from a single deck to multiple piles
Instead of having one input deck, have multiple piles for both input and
output. That is, cards can be moved between piles.

Reorient the layout so piles are to the right of the current card
details, and cards are laid out horizontally within a pile. To
accomodate this, dequeue from the front instead of the back of the pile.

Improve focus handling so the details pane more reliably receives focus.
In general, it should be focused by default and the piles pane should
not receive focus. This also lets us attach the shortcut key handler to
the details pane instead of globally.