ui: Animate cards moving between piles
Uses the View Transition API. Vue doesn't seem to have any hook we could
use to make a rendering promise, but wrapping the reactive updates seems
sufficient. nextTick() appears to be the documented way to wait on DOM
updates, but it seems to work without it.

view-transition-name should be able to take an escaped ".", but the view
transition does not trigger with it.