ui: Replace carousel scrollbars with shadows
Scroll-driven animation that makes a shadow visible when scrolled within
a carousel. Needs to be disabled with some CSS tricks[1] for
non-overflowing carousels.

Based on https://scroll-driven-animations.style/demos/scroll-shadows/css/

A simpler alternative[2] that covers the shadow with the background
color only appears to work with text the same color as the shadow.

[1]: https://brm.us/css-can-scroll
[2]: https://css-tricks.com/books/greatest-css-tricks/scroll-shadows/