ui: Workaround for card view transition escaping carousel
View transitions paint without respect to the actual page DOM[1]. This
caused artifacting if a carousel was scrolled to the right and the
overflowing cards would visibly pop in while transitioning (ironically
usually to the same position).

To workaround this we apply an instantaneous view transition on the
details pane that paints over the card view transitions. We still
observe some pop in within the piles pane over the carousel border.

Scoped view transitions[2] may be a slightly cleaner approach when they
are implemented. It would still be tricky to prevent pop in within the
piles pane, since we want cards to move across piles.

Nested view transition groups[3] may also help prevent pop in if
combined with clip-path.

Also see this blog with visual examples of the same issue[4].

Incidentally @vtbag[5] has generally useful view transition information.

[1]: https://github.com/w3c/csswg-drafts/issues/8941
[2]: https://github.com/WICG/view-transitions/blob/main/scoped-transitions.md
[3]: https://github.com/WICG/view-transitions/blob/main/nested-explainer.md
[4]: https://www.nicchan.me/blog/view-transitions-and-stacking-context/
[5]: https://vtbag.dev/