ui: Implement undo history with an operation log
A Command is a serialized form of an operation; each operation generally
has a reciprocal operation that reverses it. We can store a chain of
these in our operation log, and then invoke them in reverse to undo an
arbitrary number of operations.

The Command type indexes into a dictionary of generated types to create
a discriminated union of valid types (that is, the function name and
parameters agree). TypeScript type-level programming is seriously
impressive, and LLMs are quite helpful with it.