ui: Serialize async operations
The startViewTransition API relies on deferring state changes (so the
browser can snapshot before DOM changes). This can cause operations to
erroneously interleave when events occur too quickly.

We implement a queue for events to ensure they are executed in serial,
even if their operations are async. We use eventsPromise as an (awkward)
semaphore to ensure events are not processed concurrently.