ui: Rework card location and movement
Use Location as the main reference point for movement operations. This
helps us generalize them into a single placeCardInto function;
especially significant is uniform undo command handling, which was
previously unimplemented.

We also introduce a hidden discard pile. As an immediate benefit this
allows discard and its undo to be implemented as a simple case of card
movement. In the future the discard pile may serve as a yank buffer. We
take care to avoid user interactions with the discard pile at index 0.