app: Port api/ and ui/ to Vike
Started with scaffolding from:
$ npm create vike@latest --- --vue --ts-rest --fastify --sqlite --eslint --prettier

For a modular framework, Vike comes with a lot of opinionated
boilerplate. We remove most of it.

Uses ts-rest for the API contract. This ports the server logic from
Python and unifies it in the app server. This does not (yet) use the
contract for the client side.