collector: Reverse loop over messages
Instead of having logic to scroll to the top of the infinite scrolling
messages, just loop over messages from the bottom up.

This appears to get rid of the "Element is not attached to the DOM"
error, maybe because we're moving through messages in a more typical
order.

Collecting 349 messages takes 2:45 min, so about 2 messages/sec.