app: Virtualize CardCarousel scrolling
The precipitating need for this change was having too many elements with
view-transition-name set, which causes >1s of unresponsiveness when
moving cards. Dynamically setting view-transition-name was also
considered, but was similarly complicated; and virtualizing scrolling
also helps with overall performance.

We set the carousel scroll position directly instead of using
scrollIntoView(), so we no longer need the autoScroll event. However we
lose scrolling to the active pile, which we may want to add back.

This restores the behavior of keeping the picked card in view on
carousel resize. We also scroll to a placed card in its target pile.

Discovered these after implementing, but perhaps useful alternatives:
- https://vueuse.org/core/useVirtualList/
- https://github.com/Akryum/vue-virtual-scroller