app: Pick card that is placed into a pile
Previously the target pile retained its previous card pick. We choose to
instead pick the placed card. This probably makes more sense, but more
importantly it will be helpful with the upcoming virtualized scrolling
(to ensure the place target is not virtualized).

Virtualized scrolling will completely change how scrolling to the card
works; for now there is an artifact where we do not scroll to the placed
card in the target pile.