app: Fix card clipping artifacts during view transitions
We previously mitigated artifacting by making the details pane a fake
view transition. This still exhibited artifacts with the carousel shadow
and at the margins of the piles pane. Now we rely only on nested view
transition groups which are not yet standardized as part of CSS View
Transitions Module Level 2. We may expect more artifacts in non-Chrome
browsers.
See https://developer.chrome.com/docs/css-ui/view-transitions/nested-view-transition-groups

During view transitions, most cards are clipped to their containing
carousel; but the picked card may be moving between piles and is not
clipped. In theory we may encounter artifacting if a non-picked card
moves between piles, or if a card moves beyond the piles pane, but we do
not seem to encounter either of these as currently implemented.

Carousels and shadows are new view transition groups; the animations
subjectively feel more sluggish.