db, app: Revamp revisions and deck loading
The precipitating motivation was to make all operations infallible (they
always return a reverse). Notably we remove network operations from deck
loading; instead they are performed in the action (before calling the
operation), and we populate the OperationContext with required state
(namely the cards in a revision).

Pile state tries to be a little more careful with managing associations
with decks and revisions. In particular, we unset the revision (mark the
pile dirty) when making changes to the cards in the pile. We track these
revision association changes (that is, pile dirtiness) in the op log;
this is analogous to a text editor that knows where in the undo history
is unmodified from the file on disk.

The database now distinguishes card sets from revisions, which means we
can efficiently have multiple revisions with the same set of cards; this
may also be helpful if we want to have additional metadata on a
revision. However the hidden attribute has been removed because its
meaning under the new loading logic is unclear.

We no longer load all decks when opening the app, which avoids some
initialization races. We render deck and revision information for piles.