app: Make operations synchronous
We want to be able to replay operations instantaneously. We can only do
that if the operations can be performed synchronously. We also shouldn't
have any side effects beyond the workspace state.

We externalize view transitions and delta pushing from the operations.
There is some special casing around moveCard that might have a cleaner
solution.